+++ test/XRoadLib.Tests.Contract/ParamType1.cs
@@ -1,12 +1,19 @@
-﻿using XRoadLib.Serialization;
+﻿using XRoadLib.Attributes;
+using XRoadLib.Serialization;
 
 namespace XRoadLib.Tests.Contract
 {
     public class ParamType1 : XRoadSerializable
     {
         public long Property1 { get; set; }
+
+        [XRoadOptional]
         public ParamType2[] Property2 { get; set; }
+
+        [XRoadOptional]
         public string Property3 { get; set; }
+
+        [XRoadOptional]
         public MimeContent MimeContent { get; set; }
     }
 }